feat(shell): allow empty curly braces
The busybox shell before c470f4477ad7f548ee3ccd3662c83899a5648297 allowed empty curly braces, for example functions with no commands inside, e.g.

```
foo() {
}
```

However, this was changed in that commit. The change seems to come from the 0.4.2 release of the dash shell, since it says "banned empty compound lists in most places" in the changelog. This is part of POSIX's standard for shells, and dash tries to be compliant, so that's probably why it changed. Some firmwares use a very old version of busybox that allows empty functions, so we patch in support here. Busybox's unit tests report the same passes and fails after this change, so it probably doesn't break anything and just strictly allows more code.